Gellish Expression Syntax definition
This document defines the tabular Gellish Expression Format. The format defines the syntax of all expressions for messages, queries and databases in Gellish Formal Languages such as Formal English and Formeel Nederlands (Formal Dutch). The format can be implemented in file formats, such as CSV, XLS and JSON, but also in database formats, such as SQL.