Formal English defining ontology (basic core)
A representative core subset of the language defining ontology of Formal English that is also a taxonomic dictionary.
It defines a core set of concepts, including also a core set of kinds of relations, each with a unique identifier (UID). Their names and synonyms, base phrases and inverse phrases form a core vocabulary for making Formal English expressions.
Extensions of this taxonomic dictionary shall be proper subtypes of these concepts.
The rules for defining and extending the formal language are described in the book 'Semantic Information Modeling in Formalized Languages'.