New test data set
These files should be stored in a subdirectory Data of the directory where the program source is located.
By choosing the option 'n' (new database) while running the program, a new database will be created that imports and stores the content of these files. 
After that the content can be queried.