Example mapping table and non-Gellish JSON file
The non-Gellish JSON file is an example of query results in a database of a roof windows supplier. The file contains data in Dutch and the mapping uses data for generating a Gellish Expression Format file in Formal Dutch. The Communicator software can display the resulting model also in English, using the 'reply language' setting option.