Pickle save error corrected
The pickle save command gave an error because the file object self.f could not be pickled and was not closed.
The file object f was made local and thus closed automatically. Rewind by f.seek(0) was therefore relocated.
A db-file was generated and saved. Thus on new starts that file will be read instead of the csv files.